feat: add new metadata properties and comments
Adding a bit of comments to make the code more clear to the reader. At the same time, I thought of making the
calibration more parameterizable. Now the checkerboard size is not fixed, and the criteria for the calibration can be
configured with `calibration_accuracy` and `iterations`. Should fairly more scalable.